ctool: when collecting changes in git-export, get unstaged changes.
This changes the changes to collect the unstaged changes rather
than the staged changes.  This is interesting difference, and not sure
what the right behavior actually is.

I came to do this so that you could export after a 'git merge'
without a 'git commit'.  Without this change, the diff would
not apply as the working tree had the changes.