Do not lose proxy settings when invoking apt.
I noticed that the http_proxy settings were not getting passed through
to apt-get (run-container --install=package1), and thus the proxy
not getting used.  This was due to the apt_get function invoking
sudo.  That was not necessary as the 'install_packages' function
already ensured that it was running as root (--user=root).

So.. the easy fix is to just *not* use sudo.  It uses env instead so
that the debug statement can represent that environment var was used.