build-package: default to debuild with -nc.
If not provided with any arguments, build-package would build a source
package using '-d' (no-check-builddeps).
The change here is just to add '-nc' as well.
Without this change, the clean target will be invoked and may fail
if that target used a missing builddep.

The invocation of debuild is done in a clean git worktree in a
temporary directory, so there is no need to clean first.