ctool: launch containers with 'lxc launch ... </dev/null'
Since LXD version 3.17 the lxc command is able to read the container
configuration as YAML from stdin on `lxc init` and `lxc launch`. This
broke the ctool script, which is supposed to take a script from stdin,
but stdin is now consumed by lxc. To prevent this from happening this
commit adds a </dev/null when running `lxc launch`.

This has been reported as a regression in LXD:

  https://github.com/lxc/lxd/issues/6188