new-upstream-snapshot: fix 'is_merge_commit'
When merging  ssh-import-id from master at 5.8 to the ubuntu/debian
branch, it would cause merge conflict due to upstream maintaining a
debian/changelog.

The "fix merge commit" shell basically worked as designed, but
then when exiting the shell the 'git show --format=%P' would have changes.
We really just wanted to get the parent commits, not show diffs.

So... the fix is to just use '--no-patch'