Fix build-package and get-orig-tarball for packages with an epoch.
Packages with an epoch (leading [0-9]:) would fail to build with
build-package or to download the orig tarball with get-orig-tarball.

The crux of the issue was that both tools would leave the epoch
on the upstream version, when it should not be used.

An example that is fixed is pulseaudio.
Before, the following would fail:

    get-orig-tarball for --verbose --overwrite pulseaudio \
       1:15.0+dfsg1-1ubuntu2.1 .