new_upstream_snapshot: get previous changelog distro when UNRELEASED
This corrects the suggested dch -R -D UNRELEASED '' comment at end
of snapshot.

Also, avoid preferring previous changelog version when quilt patches
are updated AND we are snapshotting an annotated upstream tag.

The above avoids using invalid previous suffix .3 etc as we are
on a brand new XX.YY SRU so suffix should be ~18.04.1.